🐛 fix: 팔로우, 언팔로우 시 깜빡임 현상 팔로우 버튼 로직 분리로 해결 (#230)
- 팔로우, 언팔로우 API 요청 로직을 FollowButton 컴포넌트로 이동
- API 요청 로딩 시 버튼에 Spinner 컴포넌트 적용에 따라 Spinner 컴포넌트 스타일 조정
