+++ src/pages/ProfilePage/FollowPage/FollowersPage/FollowersPage.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import BasicLayout from '../../../../layout/BasicLayout';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { FollowerWrapper, FollowerList, FollowItem } from './FollowersPageStyle';
-import { useInfiniteQuery, useMutation, useQuery, useQueryClient } from 'react-query';
-import { deleteUnFollow, getFollowers, postFollow } from '../../../../api/followApi';
+import { useInfiniteQuery, useQuery } from 'react-query';
+import { getFollowers } from '../../../../api/followApi';
 import UserSimpleInfo from '../../../../components/common/UserSimpleInfo/UserSimpleInfo/UserSimpleInfo';
-import Spinner from '../../../../components/common/Spinner/Spinner';
 import useObserver from '../../../../hooks/useObserver';
 import { getMyProfile } from '../../../../api/profileApi';
 
@@ -17,7 +16,6 @@ export default function FollowersPage() {
     data: followers,
     fetchNextPage,
     isLoading,
-    isFetching,
     hasNextPage,
   } = useInfiniteQuery(
     'followers',
@@ -37,71 +35,28 @@ export default function FollowersPage() {
 
   const observerRef = useObserver(hasNextPage, fetchNextPage, isLoading);
 
-  const queryClient = useQueryClient();
-
   const { data: myProfileData } = useQuery('myProfile', getMyProfile);
-
-  // 팔로우 API
-  const postFollowMutation = useMutation(postFollow, {
-    onSuccess: () => {
-      queryClient.invalidateQueries('followers');
-    },
-    onError: () => {
-      console.error('팔로우 실패');
-    },
-  });
-
-  // 언팔로우 API
-  const deleteUnFollowMutation = useMutation(deleteUnFollow, {
-    onSuccess: () => {
-      queryClient.invalidateQueries('followers');
-    },
-    onError: () => {
-      console.error('언팔로우 실패');
-    },
-  });
 
   const handleClickBackButton = () => {
     navigate(-1);
   };
 
-  const handleClickFollow = (accountname) => {
-    // 팔로우 API 요청
-    postFollowMutation.mutate(accountname);
-  };
-
-  const handleClickUnFollow = (accountname) => {
-    // 언팔로우 API 요청
-    deleteUnFollowMutation.mutate(accountname);
-  };
-
-  if (isLoading || isFetching) {
-    return <Spinner />;
-  }
-
   return (
     <BasicLayout type='follow' title='팔로워' onClickLeftButton={handleClickBackButton}>
       <FollowerWrapper>
-        {!isLoading && (
-          <>
-            <FollowerList>
-              {followers?.map((follower) => (
-                <FollowItem key={follower._id}>
-                  <UserSimpleInfo
-                    profile={follower}
-                    type='follow'
-                    isLink={true}
-                    isMyProfile={myProfileData.accountname === follower.accountname}
-                    onClick={follower.isfollow ? handleClickUnFollow : handleClickFollow}
-                  />
-                </FollowItem>
-              ))}
-            </FollowerList>
-          </>
-        )}
-        <div ref={observerRef} style={{ minHeight: '1px' }}>
-          {(isLoading || isFetching) && <Spinner />}
-        </div>
+        <FollowerList>
+          {followers?.map((follower) => (
+            <FollowItem key={follower._id}>
+              <UserSimpleInfo
+                profile={follower}
+                type='follow'
+                isLink={true}
+                isMyProfile={myProfileData.accountname === follower.accountname}
+              />
+            </FollowItem>
+          ))}
+        </FollowerList>
+        <div ref={observerRef} style={{ minHeight: '1px' }}></div>
       </FollowerWrapper>
     </BasicLayout>
   );
